move pre-ready startup test to its own file
parallel.test.ts has several tests that exceed the 5s default
timeout under ASAN debug on constrained machines (each spawns a
coordinator + multiple workers). Isolate the pre-ready crash test
so file-level pass/fail reflects only this fix, not unrelated
timing-sensitive cases.
